arch: creating hooks to clean maps
